Refactor StarDisplay script names
