+++ Assets/_Scripts/StarPresenter.cs.meta
